requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via openai
aiosignal==1.3.2
    # via aiohttp
attrs==24.3.0
    # via aiohttp
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via nltk
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.16.1
    # via
    #   huggingface-hub
    #   transformers
fonttools==4.55.3
    # via matplotlib
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.12.0
    # via huggingface-hub
huggingface-hub==0.27.1
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   yarl
joblib==1.4.2
    # via
    #   nltk
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
matplotlib==3.10.0
    # via wordcloud
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
nltk==3.9.1
    # via textblob
numpy==2.2.1
    # via
    #   contourpy
    #   matplotlib
    #   scikit-learn
    #   scipy
    #   transformers
    #   wordcloud
openai==0.28.0
    # via -r requirements.in
packaging==24.2
    # via
    #   huggingface-hub
    #   matplotlib
    #   transformers
pillow==11.1.0
    # via
    #   matplotlib
    #   wordcloud
praw==7.8.1
    # via -r requirements.in
prawcore==2.4.0
    # via praw
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   nltk
    #   transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   openai
    #   prawcore
    #   transformers
    #   update-checker
safetensors==0.5.2
    # via transformers
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.1
    # via scikit-learn
six==1.17.0
    # via python-dateutil
textblob==0.19.0
    # via -r requirements.in
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   nltk
    #   openai
    #   transformers
transformers==4.48.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via huggingface-hub
update-checker==0.18.0
    # via praw
urllib3==2.3.0
    # via requests
websocket-client==1.8.0
    # via praw
wordcloud==1.9.4
    # via -r requirements.in
yarl==1.18.3
    # via aiohttp
